Scripts/archive/requirements.txt
appnope==0.1.0
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
bleach==2.0.0
certifi==2017.7.27.1
configparser==3.5.0
decorator==4.1.2
entrypoints==0.2.3
enum34==1.1.6
functools32==3.2.3.post2
futures==3.1.1
html5lib==0.999999999
ipykernel==4.6.1
ipython==5.4.1
ipython-genutils==0.2.0
ipywidgets==7.0.0
Jinja2==2.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.3.0
MarkupSafe==1.0
mistune==0.7.4
nbconvert==5.3.0
nbformat==4.4.0
notebook==5.0.0
pandocfilters==1.4.2
pathlib2==2.3.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
python-dateutil==2.6.1
PyYAML==3.12
pyzmq==16.0.2
qtconsole==4.3.1
requests==2.11.1
scandir==1.5
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
slacker==0.9.60
tableauserverclient==0.4.1
terminado==0.6
testpath==0.3.1
tornado==4.5.2
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.0.2
